bonus nady

+++ so_long_bonus/libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: nbouhali <nbouhali@student.1337.ma >       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 17:52:22 by nbouhali          #+#    #+#             */
-/*   Updated: 2022/10/26 02:09:04 by nbouhali         ###   ########.fr       */
+/*   Updated: 2023/02/06 22:24:31 by nbouhali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,10 +29,10 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 	while (a < len && tben[a])
 	{
 		b = 0;
-		while (tben[a + b] == lbra[b] && b < ft_strlen(needle))
+		while (tben[a + b] == lbra[b] && (int)b < ft_strlen(needle))
 		{
 			b++;
-			if (b == ft_strlen(needle) && a + b <= len)
+			if ((int)b == ft_strlen(needle) && a + b <= len)
 				return (tben + a);
 		}
 		a++;
